+++ packages/medulas-react-components/types/components/forms/TextField/index.d.ts
@@ -1,0 +1,12 @@
+import { TextFieldProps } from "@material-ui/core/TextField";
+import { FieldSubscription, FieldValidator, FormApi } from "final-form";
+import { FieldInputValue } from "../../../utils/forms/validators";
+interface InnerProps {
+  name: string;
+  form: FormApi;
+  validate?: FieldValidator<FieldInputValue>;
+  subscription?: FieldSubscription;
+}
+declare type Props = InnerProps & TextFieldProps;
+declare const TextField: ({ name, form, validate, ...restProps }: Props) => JSX.Element;
+export default TextField;
