+++ packages/sil-governance/src/routes/create-proposal/components/DescriptionField.tsx
@@ -4,7 +4,7 @@ import {
   composeValidators,
   notLongerThan,
   required,
-  TextFieldForm,
+  TextField,
   Typography,
 } from "medulas-react-components";
 import React from "react";
@@ -23,7 +23,7 @@ const DescriptionField = ({ form }: Props): JSX.Element => {
   return (
     <Block marginTop={2}>
       <Typography>{DESCRIPTION_FIELD}</Typography>
-      <TextFieldForm
+      <TextField
         name={DESCRIPTION_FIELD}
         form={form}
         validate={validator}
