+++ packages/sil-governance/src/routes/create-proposal/components/FormOptions/AmendCommitteeQuorum.tsx
@@ -4,7 +4,7 @@ import {
   FieldInputValue,
   SelectField,
   SelectFieldItem,
-  TextFieldForm,
+  TextField,
   Typography,
 } from "medulas-react-components";
 import React, { useEffect, useState } from "react";
@@ -72,7 +72,7 @@ const AmendCommitteeQuorum = ({ form }: Props): JSX.Element => {
       <Block marginTop={2} display="flex" alignItems="center">
         <Typography>{QUORUM_FIELD}</Typography>
         <Block marginLeft={2}>
-          <TextFieldForm
+          <TextField
             name={QUORUM_FIELD}
             form={form}
             validate={isFractionOrEmpty}
