+++ packages/sil-governance/src/routes/create-proposal/components/FormOptions/AmendProtocol.tsx
@@ -4,7 +4,7 @@ import {
   composeValidators,
   notLongerThan,
   required,
-  TextFieldForm,
+  TextField,
   Typography,
 } from "medulas-react-components";
 import React from "react";
@@ -23,7 +23,7 @@ const AmendProtocol = ({ form }: Props): JSX.Element => {
   return (
     <Block marginTop={2}>
       <Typography>{TEXT_FIELD}</Typography>
-      <TextFieldForm
+      <TextField
         name={TEXT_FIELD}
         form={form}
         validate={validator}
